% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tvv.R
\name{fit_tvv}
\alias{fit_tvv}
\title{Finding maximum posterior density model w/single structural matrix
(inverse(A0)) and time-varying variances for structural innovations
(exp(-lmd\link{,timePeriod}))}
\usage{
fit_tvv(
  dat,
  date_var,
  vars = NULL,
  vars_log = NULL,
  startdate = NULL,
  enddate = NULL,
  timedummy = NULL,
  period_breaks = NULL,
  n_lags = 10,
  lc_A0 = NULL,
  lmdblock = NULL,
  verbose = FALSE,
  seed_model = NULL,
  n_step = 60,
  extracheck = TRUE,
  cores = 1,
  n_varscores = 1,
  seed_x = NULL,
  lmdPrior = FALSE,
  mn_start = 1,
  mn_tight = 3,
  mn_decay = 0.5,
  v_prior = 0,
  cos_prior = NULL,
  nlt = NULL,
  hparam_nl = rep(0, 3),
  hparam = rep(0, 7),
  critval = 1e-10,
  seed_H = NULL,
  tvA = FALSE,
  noLmd = FALSE
)
}
\arguments{
\item{dat}{Matrix or data frame. Input data}

\item{date_var}{Variable in data representing date term.}

\item{vars}{Vector of variables from \emph{dat} to use in
analysis.  Defaults to all variables.}

\item{vars_log}{Vector of variables to log transform in
analysis. Defaults to all variables untransformed.}

\item{timedummy}{formatted like start/endDate, but currently not used, could
input a dummy for one/more period,}

\item{period_breaks}{Date or character vector indicating when periods end.
Should have length of (\link{# of periods to be specified} - 1).}

\item{n_lags}{Integer (or integer-like). Number of lags.}

\item{lc_A0}{An \emph{n} x \emph{n} matrix, where \emph{n} is the number of variables used.
Each element indicates TRUE for a fitted value and FALSE for a non-fitted
value. If NULL (default) uses lower triangular with TRUE on diagonal, and
zero to other FALSE fields.}

\item{lmdblock}{n_vars x nPeriods x n_blocks array that identifies, with TRUE,
blocks of parameters that are constant (presumably across periods). The
most obvious application is keeping certain parameters constant over
periods, but it could also be used to have one value in 1/2 the periods,
another in the other half, etc. If left null, all variances change in every
period.specifies exactly how to restrict some variances not to be
time-varying. Not used in the most recent calculations.}

\item{verbose}{Boolean. if TRUE (default), gives additional output during run}

\item{seed_model}{a list that looks like the output of this function from
which a first iteration x vector can be extracted intended usage: if you
tweak the prior a bit, the mode will be close to old mode downside:
estiamte of the inverse hessian from optimization will be poor}

\item{cores}{Length-1 Integer (or integer-like numeric). Number of cores used for parallelization}
}
\value{
If \code{verbose = TRUE}: list output of the optimization routine (in
$opt), plus information about the VAR estimated at the posterior mode. The
whole structure is an input for the programs that run MCMC
}
\description{
Finding maximum posterior density model w/single structural matrix
(inverse(A0)) and time-varying variances for structural innovations
(exp(-lmd\link{,timePeriod}))
}
\details{
Automatically 'cleans' but does not load the data. Need to run
'datacat' or something similar to grab data from different
sources/spreadsheets and format appropriately
}
