% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpss_tools.R
\name{gdraw}
\alias{gdraw}
\title{gibbs draws}
\usage{
gdraw(
  tvvModel,
  ndraw,
  lhfcn = bvarwrap5,
  nburn = 1000,
  nsep = 100,
  filename = "gibbsout.Rdata",
  hsnfactor = 0.2,
  alpha = 0.003,
  k = 4,
  tparam = NULL,
  tscale = 1,
  savespots = NULL,
  dscore = FALSE,
  drawdout = FALSE,
  drawe = FALSE,
  drawa = TRUE,
  hparam_nl = rep(0, 5),
  nlt = rep(0, 5)
)
}
\arguments{
\item{nlt}{}
}
\description{
gibbs draws
}
