% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpss_tools.R
\name{get_forecast}
\alias{get_forecast}
\title{gets a forecast nperiods into the future}
\usage{
get_forecast(tout, nperiods = 8, Tobs = dim(ydata)[1])
}
\arguments{
\item{tout}{full output of optimization}

\item{nperiods}{number of periods into future to forecast}

\item{Tobs}{Total number of observations}
}
\description{
gets a forecast nperiods into the future
}
