% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpss_tools.R
\name{getrmse}
\alias{getrmse}
\title{Title get root mean squared error}
\usage{
getrmse(fcast, ydata, h = c(1, 6, 12, 24, 48))
}
\arguments{
\item{fcast}{fcast array}

\item{ydata}{model data for Y}

\item{h}{}
}
\description{
Title get root mean squared error
}
