% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{impulseplots}
\alias{impulseplots}
\title{Impluse plots}
\usage{
impulseplots(
  ir,
  irdraws = NULL,
  conf = c(0.68, 0.9),
  blocks = NULL,
  filename = "impulse_plot",
  format = "pdf",
  addtitle = FALSE,
  nsteps = 60,
  shocknames = NULL,
  varnames = rep("", dim(ir)[1]),
  ptype = "Impulse response",
  color = c(0, 0, 1),
  alpha = c(0.5, 0.3),
  gr = 0.7,
  width = 5,
  height = 8,
  response_ylim = NULL,
  xtick = NULL
)
}
\arguments{
\item{xtick}{}
}
\description{
Impluse plots
}
