% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpss_tools.R
\name{McmcIr}
\alias{McmcIr}
\title{Main wrapper function for drawing impulse responses}
\usage{
McmcIr(
  xout,
  model,
  rootcheck = FALSE,
  lrange = 1:6,
  cores = 8,
  oweights = NULL,
  aplus = NULL,
  nStep = 60,
  hparam = rep(0, 7)
)
}
\arguments{
\item{xout}{A0 and Lambda draws in matrix form}

\item{rootcheck}{if TRUE, will check if roots are stable}

\item{lrange}{which variance regimes you want IR for. The IR are all the same up to scale}

\item{cores}{number of cores to use for mclapply. This is easily parallelized, so use many!}

\item{oweights}{draws of delta_it}

\item{aplus}{draws of A+. Code will run a lot faster if you specify this}

\item{nStep}{number of steps of impulse responses}

\item{hparam}{}
}
\value{
list with element ir, the impulse responses \link{nvar x nshock x nperiod x nregime x ndraws}, plus other stuff which is less useful (mainly related to orderings/permutations)
}
\description{
Main wrapper function for drawing impulse responses
}
