% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_tools.R
\name{mgnldnsty}
\alias{mgnldnsty}
\title{Title}
\usage{
mgnldnsty(
  ydata,
  lags,
  xdata = NULL,
  const = TRUE,
  breaks = NULL,
  lambda = 5,
  mu = 1,
  mnprior = list(tight = 3, decay = 0.5),
  vprior = list(sig = NULL, w = 1),
  train = 0,
  flat = FALSE,
  nonorm = FALSE,
  ic = NULL
)
}
\arguments{
\item{ic}{}
}
\description{
Title
}
