% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_model}
\alias{sim_model}
\title{Title}
\usage{
sim_model(
  tvvout,
  lambda = matrix(1, nv, 6),
  lmdp = rep(10, 6),
  burn = 0,
  y0 = c(rep(0, nv * nlag), 1),
  delta = NULL
)
}
\arguments{
\item{delta}{}
}
\description{
Title
}
