% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpss_tools.R
\name{plotfc}
\alias{plotfc}
\title{Title forecast plotting}
\usage{
plotfc(
  fcout,
  ydata,
  dateseq,
  vnames,
  fulldates,
  ymat = c(4.35, 4.7, 4.55, 4.65, 8.2, 8.5, 7, 7.4, 7.2, 7.6, -0.02, 0.06, 5.6, 6.4,
    -0.01, 0.05, 0, 0.12, -0.01, 0.05),
  filename = "fcout",
  cushion = 0
)
}
\arguments{
\item{fcout}{forecasting results}

\item{cushion}{}
}
\description{
Title forecast plotting
}
