% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotrmse}
\alias{plotrmse}
\title{generate rmse plots}
\usage{
plotrmse(
  sdate,
  rmse1,
  rmse2,
  rmse3,
  filename,
  vname = rep(NULL, nv),
  ih = 1,
  height = 8,
  width = 6.5,
  diff = FALSE,
  rshade = TRUE
)
}
\arguments{
\item{rshade}{}
}
\description{
generate rmse plots
}
